requirements.txt
Flask==2.3.3
cryptography==41.0.7
PyMySQL==1.1.0
PyYAML==6.0.1
psutil==5.9.5
gunicorn==21.2.0
python-dotenv==1.0.0 